Eager-bound-supertraits are a nice pattern
I've changed PipeStream to require IntoRawHandle and FromRawHandle in
addition to AsRawHandle and NamedPipeStreamInternals. What this does is
that whenever a trait bound uses the PipeStream trait, it will
automatically pull in IntoRawHandle and FromRawHandle, decreasing typing
redundancy if those need to be bounded as well. This pattern is known as
"eager bound supertraits", because it's about specifying all traits
known to be implemented on all implementations of a sealed trait. Please
follow this pattern when contributing, it adds a lot to API ergonomics.