Replaced upgradable reads with write relocking
Upgradable reads on RW-locks actually prevent additional read locks,
i.e. if all guards on an RW-lock are either upgradable reads or writes,
it will act like a regular mutex but with the overhead of an RW-lock.
Dropping the read guard and locking with write access obviously benefits
the readers and disadvantages the writer, but since most accesses are
supposed to be reads anyway, that's exactly what we want.