Soundness fixes for ancillary data and use of `readv`/`writev`
My dumb ass didn't know what alignment is back when I was writing this
code, so coming back to this crate to see a basically guaranteed
`SIGBUS` nearly everywhere other than x86 was a "pleasant" surprise.
This commit disables any and all usage of ancillary data on non-x86
architectures, which shouldn't have a *massive* impact since IPC is not
as widely used on embedded non-desktop platforms, so the only people who
are really going to notice this are those who use software written in
Rust which is "bulky" enough to warrant IPC, on a  Raspberry Pi
workstation. If they choose to update `interprocess`, that is. I'm
releasing version 2.0.0 which fixes this and restores the functionality
simultaneously with 1.2.0, so upgrading should be possible immediately.