Removed system calls from named pipe local socket wrapper
Since the `from_raw_handle()` constructor now panics if it can't tell if
the handle belongs to a client or a server, and the underlying named
pipe knows that right off the bat, there's no need for the Nah variant
(and no need for that whole dumb struct to begin with).
`os::windows::local_socket::LocalSocketStream` now only has one field.