Remove `to_method` dependency
The `to_method` crate is licensed with CC0. The CC0 license is
primarily intended for non-software work and is not recommended
for software works due to the explicit lack of patent grants.

The `to_method` crate only adds suffix notation for conversion
functions.

This commit removes the dependency and uses the standard prefix
notations instead.

https://www.gnu.org/licenses/license-list.en.html#CC0