chore: relicense as 0BSD
I have long ceased to believe that intellectual property as enforced by
the law is a greater force for software freedom than it is an impediment
to developers who are just trying to get by in this increasingly
tiresome world we live in.

Recently, I decided to interrogate the usefulness of software licenses
that seek to apply the state's monopoly on violence to bring about
something as trivial as acknowledging the authorship of upstream
software. A fact of life that software freedom ideologues of the less
grounded-in-material-reality sort like to overlook is that, while
enforcement of copyright law is equal (enabling the concept of copyleft
and things of that nature), it is anything but equitable. Accordingly,
the worst potential offenders of the social norms that permissive
software licenses try to enforce have ample resources to intimidate
individuals into not taking legal action against license violations,
defeat those who are not easily intimidated in an unfair battle of
top-rate legal defense against lawyers less than conversational in the
legal alternative reality conceived by tech companies, and simply lobby
the rule of law out of existence when all else fails.

The bottom line, which is evident even if you firmly believe that
proprietary products should obligatorily disclose every piece of
open-source software they use, is that no one in their right mind would
ever sue anyone for a lack of a shoutout in an OSS credits list, and as
such, I've come to the conclusion that I cannot justify preferring `MIT
OR Apache-2.0` over the simpler and less invasive 0BSD.

For those using Interprocess in open-source Rust crates that pull it in
as a dependency and do not vendor it, nothing has changed whatsoever.
Those who vendor Interprocess or redistribute it as part of compiled
binary programs without available source code will no longer be
obligated by law to mention that Interprocess is used by the program,
although I would much appreciate if you did that regardless: now that
acknowledging my work on Interprocess is no longer mandatory, you can do
so with recognizable respect and sincerity.

Say "no" to licenses and "yes" to social norms.